kafka-mirrormaker2/migrate-dedicated-to-connect/requirements.txt
kafka-python>=2.2.15